+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-2.1.0 - Unreleased
+2.1.0 - 2026-05-08
 ------------------
 
 - With devpi-server 7.0.0 filtering when inheriting from a constrained index works as expected.